Split up pages
Used hash/state based navigation, but I'm not a fan, a lot of finagling
to get working what a simple folder structure does for free

There is some copy pasting going on, it was either that or setup some
templating ala handlebars, I opted for the former, to keep things simple
