+++ src/community/index.html
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta name="viewport" content="width=device-width">
+    <link rel="stylesheet" href="../style.css">
+    <link rel="icon" type="image/png" href="../favicon.png" />
+    <meta name="description" content="Homepage to the srcery syntax theme">
+    <title>Srcery Theme - Syntax theme with clearly defined contrasting colors and a slightly earthy tone</title>
+  </head>
+  <body class="font-iosevka bg-black text-bright_white flex justify-center p-4 w-100">
+    <main x-data="state" class="flex items-center flex-col grow sm:grow-0 width-[448px] sm:w-[780px] min-w-0">
+      <header class="min-w-auto sm:min-w-[640px] sm:max-w-[640px] mb-8">
+        <a href="/" class="min-h-auto pt-2 sm:pt-0 sm:min-h-[546px] flex flex-col items-center mb-2">
+          <div class="flex sm:hidden p-5 rounded-full border-2 border-bright_black justify-center items-center">
+            <img alt="srcery glyph" class="" src="../svg/glyph.svg" class="glyph w-logo h-logo">
+          </div>
+          <div class="bg-triangle bg-no-repeat min-h-[546px] pt-[30%] hidden sm:flex min-w-full bg-center align-center justify-center">
+            <img alt="srcery glyph" src="../svg/glyph.svg" class="glyph w-logo h-logo">
+          </div>
+          <div class="mt-4 sm:mt-[-18%] flex items-center">
+            <div class="flex mr-2">
+              <div class="dot bg-xgray2"></div>
+              <div class="dot bg-xgray7"></div>
+              <div class="dot bg-xgray10"></div>
+            </div>
+            <h1 class="uppercase text-6xl">srcery</h1>
+            <div class="flex ml-2">
+              <div class="dot bg-xgray10"></div>
+              <div class="dot bg-xgray7"></div>
+              <div class="dot bg-xgray2"></div>
+            </div>
+          </div>
+        </a>
+        <nav class="flex justify-center max-w-[384px] sm:max-w-full flex grow justify-between">
+          <span class="w-full text-left">
+            <a href="/themes/" class="link uppercase">themes</a>
+          </span>
+          <span class="w-full text-center">
+            <a href="/resources/" class="link uppercase">resources</a>
+          </span>
+          <span class="w-full text-right">
+            <a href="/community/" class="link uppercase">community</a>
+          </span>
+        </nav>
+      </header>
+      <section class="w-full">
+        hello community
+      </section>
+    </main>
+    <script type="module" src="main.js"></script>
+  </body>
+</html>
