+++ src/community/main.js
@@ -1,0 +1,5 @@
+function main() {
+  console.log("hello community")
+}
+
+main();
